fix serialize bug, change valid locate and fix dtype

+++ tfdet/core/anchor/generator.py
@@ -1,7 +1,7 @@
 import tensorflow as tf
 import numpy as np
 
-def generate_anchors(feature, image_shape = [1024, 1024], scale = [0.03125, 0.0625, 0.125, 0.25, 0.5], ratio = [0.5, 1, 2], normalize = True, auto_scale = True, flatten = True, concat = True):
+def generate_anchors(feature, image_shape = [1024, 1024], scale = [0.03125, 0.0625, 0.125, 0.25, 0.5], ratio = [0.5, 1, 2], normalize = True, auto_scale = True, flatten = True, concat = True, dtype = tf.float32):
     """
     feature = feature or [features] or shape or [shapes]
     scale = anchor_size scale
@@ -41,13 +41,13 @@ def generate_anchors(feature, image_shape = [1024, 1024], scale = [0.03125, 0.06
                 ndim = (tf.keras.backend.ndim(shape) if tf.is_tensor(shape) else np.ndim(shape)) - np.ndim(scale)
                 for _ in range(ndim):
                     shape = tf.expand_dims(shape, axis = -1)
-                scale = tf.divide(tf.cast(scale, tf.float32), tf.cast(shape, tf.float32))
+                scale = tf.divide(tf.cast(scale, dtype), tf.cast(shape, dtype))
             else:
                 stride = image_shape
         
         # Generate base anchor
-        scale = tf.cast(tf.expand_dims(scale, axis = 1), tf.float32)
-        ratio = tf.cast(tf.expand_dims(ratio, axis = 1), tf.float32)
+        scale = tf.cast(tf.expand_dims(scale, axis = 1), dtype)
+        ratio = tf.cast(tf.expand_dims(ratio, axis = 1), dtype)
 
         h = tf.reshape(scale[0] * tf.sqrt(ratio), (-1, 1))
         w = tf.reshape(scale[1] / tf.sqrt(ratio), (-1, 1))
@@ -55,15 +55,15 @@ def generate_anchors(feature, image_shape = [1024, 1024], scale = [0.03125, 0.06
         anchors = tf.concat([-0.5 * w, -0.5 * h, 0.5 * w, 0.5 * h], axis = -1)
 
         # Shift base anchor
-        center_x = (tf.cast(tf.range(feature_shape[1]), tf.float32) + 0.5) * tf.cast(1 / feature_shape[1], tf.float32) * tf.cast(stride[1], tf.float32)
-        center_y = (tf.cast(tf.range(feature_shape[0]), tf.float32) + 0.5) * tf.cast(1 / feature_shape[0], tf.float32) * tf.cast(stride[0], tf.float32)
+        center_x = (tf.cast(tf.range(feature_shape[1]), dtype) + 0.5) * tf.cast(1 / feature_shape[1], dtype) * tf.cast(stride[1], dtype)
+        center_y = (tf.cast(tf.range(feature_shape[0]), dtype) + 0.5) * tf.cast(1 / feature_shape[0], dtype) * tf.cast(stride[0], dtype)
 
         center_x, center_y = tf.meshgrid(center_x, center_y)
         center_x = tf.reshape(center_x, [-1])
         center_y = tf.reshape(center_y, [-1])
 
         shifts = tf.expand_dims(tf.stack([center_x, center_y, center_x, center_y], axis = 1), axis = 1)
-        anchors = tf.cast(tf.expand_dims(anchors, axis = 0), tf.float32) + shifts
+        anchors = tf.cast(tf.expand_dims(anchors, axis = 0), dtype) + shifts
         shape = [-1, 4]
         if not flatten:
             shape = tf.concat([feature_shape, shape], axis = 0)
@@ -78,7 +78,7 @@ def generate_anchors(feature, image_shape = [1024, 1024], scale = [0.03125, 0.06
 def generate_yolo_anchors(feature, image_shape = [608, 608], size = [[0.01645, 0.02138], [0.02632, 0.04934], [0.05428, 0.03783],
                                                                      [0.04934, 0.10033], [0.10197, 0.07401], [0.09704, 0.19572],
                                                                      [0.19079, 0.14803], [0.25658, 0.32566], [0.61349, 0.53618]], 
-                          normalize = True, auto_size = True, flatten = True, concat = True):
+                          normalize = True, auto_size = True, flatten = True, concat = True, dtype = tf.float32):
     """
     feature = feature or [features] or shape or [shapes]
     size = anchor_size (w, h)
@@ -114,27 +114,27 @@ def generate_yolo_anchors(feature, image_shape = [608, 608], size = [[0.01645, 0
                 ndim = (tf.keras.backend.ndim(shape) if tf.is_tensor(shape) else np.ndim(shape)) - np.ndim(size)
                 for _ in range(ndim):
                     shape = tf.expand_dims(shape, axis = -1)
-                size = tf.divide(tf.cast(size, tf.float32), tf.cast(shape, tf.float32))
+                size = tf.divide(tf.cast(size, dtype), tf.cast(shape, dtype))
             else:
                 stride = image_shape
         
         # Generate base anchor
-        w, h = tf.split(tf.cast(size, tf.float32), 2, axis = -1)
+        w, h = tf.split(tf.cast(size, dtype), 2, axis = -1)
         w = tf.reshape(w, [-1, 1])
         h = tf.reshape(h, [-1, 1])
 
         anchors = tf.concat([-0.5 * w, -0.5 * h, 0.5 * w, 0.5 * h], axis = -1)
 
         # Shift base anchor
-        center_x = (tf.cast(tf.range(feature_shape[1]), tf.float32) + 0.5) * tf.cast(1 / feature_shape[1], tf.float32) * tf.cast(stride[1], tf.float32)
-        center_y = (tf.cast(tf.range(feature_shape[0]), tf.float32) + 0.5) * tf.cast(1 / feature_shape[0], tf.float32) * tf.cast(stride[0], tf.float32)
+        center_x = (tf.cast(tf.range(feature_shape[1]), dtype) + 0.5) * tf.cast(1 / feature_shape[1], dtype) * tf.cast(stride[1], dtype)
+        center_y = (tf.cast(tf.range(feature_shape[0]), dtype) + 0.5) * tf.cast(1 / feature_shape[0], dtype) * tf.cast(stride[0], dtype)
 
         center_x, center_y = tf.meshgrid(center_x, center_y)
         center_x = tf.reshape(center_x, [-1])
         center_y = tf.reshape(center_y, [-1])
 
         shifts = tf.expand_dims(tf.stack([center_x, center_y, center_x, center_y], axis = 1), axis = 1)
-        anchors = tf.cast(tf.expand_dims(anchors, axis = 0), tf.float32) + shifts
+        anchors = tf.cast(tf.expand_dims(anchors, axis = 0), dtype) + shifts
         shape = [-1, 4]
         if not flatten:
             shape = tf.concat([feature_shape, shape], axis = 0)
@@ -162,16 +162,16 @@ def generate_points(feature, image_shape = [1024, 1024], stride = None, normaliz
         
         stride = [stride, stride]
         if stride[0] is None:
-            stride = tf.divide(tf.cast(image_shape, tf.float32), tf.cast(shape, tf.float32))
+            stride = tf.divide(tf.cast(image_shape, dtype), tf.cast(shape, dtype))
             if normalize:
-                stride = tf.where(tf.greater(stride, 1), tf.divide(stride, tf.cast(image_shape, tf.float32)), stride)
+                stride = tf.where(tf.greater(stride, 1), tf.divide(stride, tf.cast(image_shape, dtype)), stride)
         elif normalize and 1 < tf.reduce_max(stride):
-            stride = tf.divide(stride, tf.cast(image_shape, tf.float32))
-        shift = tf.cast(tf.divide(stride, 2), tf.float32)
+            stride = tf.divide(stride, tf.cast(image_shape, dtype))
+        shift = tf.cast(tf.divide(stride, 2), dtype)
         shift = tf.where(tf.greater(stride[0], 1), tf.floor(shift), shift)
         
         x, y = tf.meshgrid(tf.range(shape[0]), tf.range(shape[1]))
-        point = tf.cast(tf.stack([x, y], axis = -1), tf.float32) * stride[::-1] + shift[::-1]
+        point = tf.cast(tf.stack([x, y], axis = -1), dtype) * stride[::-1] + shift[::-1]
         if flatten:
             point = tf.reshape(point, [-1, 2])
         out.append(tf.stop_gradient(point))
